Restructure codebase to be more navigable and less daunting for new users.
Prior to this change the codebase was functionally structured around the Puppet code. As we use a non-standard way of deploying Puppet here this directory structure was not based on any common convention.

This change introduces a new directory layout, that also doesn't follow any common standard I know of, but which should give 'users' a better navigable codebase where priorities are separated and complexity is hidden.

'users' are considered individuals that use this repo to deploy a Failmap instance, but are not really involved otherwise with the project and thus won't develop that much or have/require a good understanding of the deployment codebase.

Functionally nothing really changed.